🐛[fix]: 코맨트 반영 및 Merge 하면서 누락된 부분 추가
- ArchiveView에서 상단 부분 블러처리 안되는 것 수정
- 기타 코맨트 반영완료
