+++ Projects/LastDance/LastDance.xcodeproj/xcshareddata/xcschemes/LastDance.xcscheme
@@ -52,12 +52,12 @@
       </BuildableProductRunnable>
       <AdditionalOptions>
          <AdditionalOption
-            key = "MallocScribble"
+            key = "MallocGuardEdges"
             value = ""
             isEnabled = "YES">
          </AdditionalOption>
          <AdditionalOption
-            key = "MallocGuardEdges"
+            key = "MallocScribble"
             value = ""
             isEnabled = "YES">
          </AdditionalOption>
