+++ Projects/LastDance/LastDance/Sources/Presentation/Reaction/Views/TagSelectView.swift
@@ -57,7 +57,6 @@ struct TagSelectView: View {
                     return false
                 }
             }
-            .padding(.bottom, 10)
         }
         .background(Color(red: 0.95, green: 0.95, blue: 0.95))
         .navigationBarTitleDisplayMode(.inline)
